fix: create verifying blockstore to avoid ts ignores
