+++ src/unpack/index.ts
@@ -1,34 +1,24 @@
-import { equals } from 'uint8arrays'
-import { sha256 } from 'multiformats/hashes/sha2'
 import toIterable from 'browser-readablestream-to-it'
 import { CarBlockIterator } from '@ipld/car/iterator'
 import { CarReader } from '@ipld/car/api'
-import { Block } from '@ipld/car/api'
 import { CID } from 'multiformats'
 import exporter from 'ipfs-unixfs-exporter'
 import type { UnixFSEntry } from 'ipfs-unixfs-exporter'
 export type { UnixFSEntry }
 
+import { VerifyingGetOnlyBlockStore } from './utils/verifying-get-only-blockstore'
 import { Blockstore } from '../blockstore/index'
 import { MemoryBlockStore } from '../blockstore/memory'
 
 // Export unixfs entries from car file
 export async function* unpack(carReader: CarReader, roots?: CID[]): AsyncIterable<UnixFSEntry> {
-  const verifyingBlockService = asVerifyingGetOnlyBlockStore({
-    // Return bytes in the same fashion as a Blockstore implementation
-    get: async (cid: CID) => {
-      const block = await carReader.get(cid)
-
-      return block?.bytes
-    }
-  })
+  const verifyingBlockService = VerifyingGetOnlyBlockStore.fromCarReader(carReader)
 
   if (!roots || roots.length === 0 ) {
     roots = await carReader.getRoots()
   }
 
   for (const root of roots) {
-    // @ts-ignore blockstore type is a superset of the verify block service needed
     yield* exporter.recursive(root, verifyingBlockService, { /* options */ })
   }
 }
@@ -41,21 +31,15 @@ export async function* unpackStream(readable: ReadableStream<Uint8Array> | Async
     await blockstore.put(block.cid, block.bytes)
   }
 
-  const verifyingBlockStore = asVerifyingGetOnlyBlockStore(blockstore)
+  const verifyingBlockStore = VerifyingGetOnlyBlockStore.fromBlockstore(blockstore)
 
   if (!roots || roots.length === 0 ) {
     roots = await carIterator.getRoots()
   }
 
   for (const root of roots) {
-    // @ts-ignore blockstore type is a superset of the verify block service needed
     yield* exporter.recursive(root, verifyingBlockStore)
   }
-}
-
-async function isValid ({ cid, bytes }: Block) {
-  const hash = await sha256.digest(bytes)
-  return equals(hash.digest, cid.multihash.digest)
 }
 
 /**
@@ -69,20 +53,3 @@ async function isValid ({ cid, bytes }: Block) {
    return Symbol.asyncIterator in readable ? readable : toIterable(readable)
 }
 
-function asVerifyingGetOnlyBlockStore(blockstore: { get: (cid: CID) => Promise<Uint8Array|undefined> }) {
-  return {
-    get: async (cid: CID) => {
-      const res = await blockstore.get(cid)
-      if (!res) {
-        throw new Error(`Incomplete CAR. Block missing for CID ${cid}`)
-      }
-      if (!isValid({ cid, bytes: res })) {
-        throw new Error(`Invalid CAR. Hash of block data does not match CID ${cid}`)
-      }
-      return res
-    },
-    put: ({ cid, bytes }: { cid: CID, bytes: Uint8Array }) => {
-      return Promise.reject(new Error('should not put blocks'))
-    }
-  }
-}
